last phase tester

+++ ft_printf/ft_convert_hex_ptr.c
@@ -6,15 +6,15 @@
 /*   By: pgiraude <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/19 19:49:46 by pgiraude          #+#    #+#             */
-/*   Updated: 2022/08/19 20:21:21 by pgiraude         ###   ########.fr       */
+/*   Updated: 2022/08/19 21:01:34 by pgiraude         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_convert_hex_ptr(int flag)
+int	ft_convert_hex_ptr(unsigned long int flag)
 {
-	int	len;
+	unsigned long int	len;
 
 	len = 0;
 	if (flag)
