+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: pgiraude <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/19 19:34:19 by pgiraude          #+#    #+#             */
-/*   Updated: 2022/08/19 20:40:12 by pgiraude         ###   ########.fr       */
+/*   Updated: 2022/08/19 21:34:51 by pgiraude         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@
 # include "libft/libft.h"
 
 int	ft_printf(const char *str, ...);
-int	ft_convert_hex_len(int flag, int mode);
-int	ft_convert_hex_ptr(int flag);
-int	ft_nbrconvert_len(int nbr, int base);
+int	ft_convert_hex_len(unsigned long int flag, int mode);
+int	ft_convert_hex_ptr(unsigned long int flag);
+int	ft_nbrconvert_len(long int nbr, int base);
 int	ft_putnbr_printf(int flag);
 int	ft_putstr_printf(char *str);
 int	ft_uputnbr_printf(unsigned int flag);
